+++ middleware.ts
@@ -1,0 +1,11 @@
+import { NextResponse } from "next/server";
+import type { NextRequest } from "next/server";
+
+export function middleware(request: NextRequest) {
+  console.log("Middleware entered! Path:", request.nextUrl.pathname);
+  return NextResponse.next();
+}
+
+export const config = {
+  matcher: ['/:path*'],
+};
